pacman: added a progress bar which is used by both pacman and spynnaker
added changes to spynnaker so that it should be close to completion. It now has the code for running and generating. Likely buggy though.

STILL REQUIRES PACMAN functions to be debuggable

added a optimal parameter to spinnman's app signal code (count needs more data and this shows that the app signal stuff hasnt been fully impliemnted so far)