added the subedge to the routing info object (didnt relaise it needed one becuase o fthe named tuple thing).
ive also completed the push of the router check for cycles and that we actually reach the correct places as a check that can be called with in debug mode

a mode has been added to the cfg. Attempts were made to use the logger mode to do this, but this failed as it seems we dont set the elvel and even whenw e do, the logger does not give a stable way of getting if it is set to debug mode.