load of wee changes to get this branch stableish.
so fixed database to deal with new key_and_mask stuff.

fixed missing names or typos which would cause system failure.

fixed a few bugs in the malloc key allocator

fixed a load of bugs in external device to make it close to stable

NOTE: retinas currently do not work. The changes needed to get these to wrok will be in the next few commits.