started the multi-run branch.
each population has a changed flag which can be set true or false to deduce if things have changed since it was last looked at.

considerations for when the graph has not been changed are now within the front end commona nd spinnaker.py.

If the runtime is bigger than last time, it goes through mapping, if things have changed it goes though mapping.

Cores now get two new sdp messages, one is for runtime, and one is to exit.

considerations have to be done for get spikes, get gsyn, get v, as pynn states get_spikes needs to return the data for all runs.

reload needs to be adjusted to reflect the switch to syno0 and syno1

c code needs modifying as stated.

app stop's unrelaibility means we avoid using it with the sdp.