Add default value to SpikeSourceArray spike_times
SpikeSourceArray did not specify a default value for spike_times
in its constructor. This caused sPyNNaker to crash with the
weird exception

    File "sPyNNaker/spynnaker/pyNN/models/pynn_population.py", line 79, in __init__
      self._vertex = cellclass(**cellparams)
    TypeError: __init__() takes at least 5 arguments (5 given)

if an empty parameter dictionary was passed to the Population
constructor (which supposedly should not happen):

    sim.Population(10, sim.SpikeSourceArray, {})

This fix simply adds the missing default value to the spike_times
parameter. Note, that the parameter order in the
SpikeSourceArray constructor was altered.