tests, changes etc for pop, pop views, and assembliers. Currently all the interface level stuff works. Most of auto pause and resume works (apart from adding a new populuation after a run with reset.
and SSP still needs to be fixed, as well as external devices, and there's a issue with synaptic matrix's somewhere. But the back of the work is done now.

The code needs a bit of a refactor to be honest, but thats left till got a fully working code base to refactor.