Add punishment signal
In order to add punishment signal, I separate 'supervision' synapses into two
different target types: reward and punishment. When synapse is punishment type,
dopamine concentration that dopaminergic neuron brings is inverted to cause
decrease in weight. This is not biologically plausible and other schemes exist,
e.g. having a base level of dopamine above which it is reward and below is
punishment.