First changes to loop interior of do_timestep_update
Not happy yet with how this is implemented within the header files
in implementations/ and input_types/; looking to do something with
a struct within neuron_impl*.h but haven't worked it out yet.  A
lot of what is commented out in neuron/neuron.c is an attempt to
make this work.