merged roberts from_list_edge_filter branch into atom_routing branch
stripped out a lot of the python so that it doesnt expand synamptic matrix.
so synaptic_matrix, master pop tables, projection_machine_edge and from_list_connector is now matched to master.

stripped out a bit in the c code. so neuron-typedefs.h is now matching master.

stripped out a bunch of c code so that its read from a seperate region and
has a differnt format of
n bitfields, key, x words, bitfield. rinse and repeat

added new region and sdram calcs for it in APV.
added provenance data items accordingly (pop machine vertex)

moved the master pop search a bit, so that i can call it in init for the bitfields
so that it can map to the master pop entry properly.

added comments and c doc to the roberts code that is being kept.