got somewhere.
at least. the python should now be in a state where it will run the machine level compressor and if the sdram mallocing fails for any reason, itll go to the host router compressor and regenerate the synaptic matrix regardless.

the c code has a bit done, its got the key to atom read done, its currently doing the read in bitfield maps and has the n_redundant packets processing complete.

the c code has not even sniffed a compiler yet.

there's a bunch of structs used to try to encapsulate the dicts and lists etc.

the fake heap has been hooked, but not implemented, this needs to be done.
the binary search has not been done, that needs to be implmented.
the cleaning of the ordered covering to not explode has not been done, this needs to be

fiddles wih expresso to make it more usable for where the expresso exe is has been done (pointless work i think)

python api been split up so that you dont nee dto have a synaptic matrix to be able to utilise the bitfield router compressor

synaptic manager has been adjusted so that it should put all the expanded stuff at the bottom of a sdram point, which is handed to the router compressor loader for sdram exploitation.