The fixed impl of the pop table is broken. Remove it!
It hasn't actually been safe to call from the rest of the C code for
quite a while; even the initialisation function's formal arguments have
diverged. Since there's no way it can be safely used, and the effort to
restore it is unreasonable, it's best we just get rid; unsupported is
simply unsupported.