Magic, with __builtin_assume_aligned()!
This prevents GCC from deciding that it has to convert a simple copy of
a four-byte structure into a call to memcpy(), a function that is
hundreds of bytes long on ARM. This is the only place in our code where
this was happening, and it was all because the compiler was thinking
"but _what if_ this field happens to have a non-aligned value put in it,
hmm?" when we can outright prove ourselves that this never happens.

The portable(-ish) syntax for doing this is horrible.