There can be only one...
... use of a random number source per selected random partner. Note that
this works because the original selected neuron ID just needs to be
converted into a subpop/local ID pair. We simply don't need to draw
another random number to achieve the choice of the second part; there
was already sufficient random information in the first random number to
provide an adequate selection. We just need to stop throwing the
information away!